chore: more reusable <RegistriesMetadata>
