+++ lib/registries/addon/components/gutters/styles.scss
@@ -34,11 +34,12 @@
         position: absolute;
         #{$side}: 0; // Fix side to the edges of the parent element
         height: 100%;
+        overflow-y: auto;
         z-index: 20;
 
         // Display a box shadow to either side when opened
         &[data-gutter-closed='false'] {
-            box-shadow: ($modifier * 5px) 0 5px -4px rgba(0, 0, 0, 0.2);
+            box-shadow: ($modifier * -5px) 0 5px -4px rgba(0, 0, 0, 0.2);
         }
 
         // Slide off the page when closed
