+++ lib/registries/addon/components/registries-metadata/component.ts
@@ -1,22 +1,34 @@
 import { tagName } from '@ember-decorators/component';
 import { computed } from '@ember-decorators/object';
 import Component from '@ember/component';
+import config from 'ember-get-config';
 
 import { layout } from 'ember-osf-web/decorators/component';
 import Registration from 'ember-osf-web/models/registration';
-import { GutterMode } from 'registries/components/gutters/component';
 import styles from './styles';
 import template from './template';
+
+const {
+    OSF: {
+        url: osfUrl,
+    },
+} = config;
 
 @tagName('')
 @layout(template, styles)
 export default class RegistriesMetadata extends Component {
-    gutterMode!: GutterMode;
-
     registration?: Registration;
 
-    @computed('gutterMode')
-    get isPage() {
-        return this.gutterMode === 'page';
+    @computed('registration')
+    get registeredFromId() {
+        if (!this.registration) {
+            return null;
+        }
+        return this.registration.belongsTo('registeredFrom').id();
+    }
+
+    @computed('registeredFromId')
+    get registeredFromDisplayUrl() {
+        return `${osfUrl.split('//')[1]}${this.registeredFromId}`;
     }
 }
